refactor: move the ui components to the correct package and delete not used files and code.
