+++ ui/src/main/java/com/amsterdam/ui/components/movieAndTvShowDetails/AddToListDialog.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.ui.components
+package com.amsterdam.ui.components.movieAndTvShowDetails
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
