+++ ui/src/main/java/com/amsterdam/ui/screens/categories/CategoriesScreen.kt
@@ -33,7 +33,7 @@ import com.amsterdam.designsystem.components.TabsLayout
 import com.amsterdam.designsystem.components.Text
 import com.amsterdam.designsystem.theme.AppTheme
 import com.amsterdam.ui.application.LocalNavManager
-import com.amsterdam.ui.components.CategoryCard
+import com.amsterdam.ui.screens.categories.components.CategoryCard
 import com.amsterdam.ui.screens.categoriesDetails.GenreMovieUiModel
 import com.amsterdam.ui.screens.categoriesDetails.GenreTvShowUiModel
 import com.amsterdam.viewmodel.categories.CategoriesInteractionListener
