+++ ui/src/main/java/com/amsterdam/ui/screens/profile/components/ProfileImagePlaceholder.kt
@@ -1,11 +1,10 @@
-package com.amsterdam.ui.components
+package com.amsterdam.ui.screens.profile.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import com.amsterdam.ui.R
-
 @Composable
 fun ProfileImagePlaceholder() {
     Image(
