+++ ui/src/main/java/com/amsterdam/ui/screens/profile/components/ThemeSelectionItem.kt
@@ -1,5 +1,7 @@
-package com.amsterdam.ui.components.selection
+package com.amsterdam.ui.screens.profile.components
 
+import androidx.annotation.DrawableRes
+import androidx.compose.animation.animateColorAsState
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -13,10 +15,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.amsterdam.ui.R
 import com.amsterdam.designsystem.components.Icon
 import com.amsterdam.designsystem.components.buttons.RadioButton
 import com.amsterdam.designsystem.components.buttons.RadioState
@@ -24,35 +26,48 @@ import com.amsterdam.designsystem.components.Text
 import com.amsterdam.designsystem.theme.AflamiTheme
 import com.amsterdam.designsystem.theme.AppTheme
 import com.amsterdam.designsystem.utils.ThemeAndLocalePreviews
+import com.amsterdam.ui.R
 
 @Composable
-fun AnswerSelectionItem(
+fun ThemeSelectionItem(
+    @DrawableRes trailingIcon: Int,
     modifier: Modifier = Modifier,
     text: String = "",
-    isClickable : Boolean = true,
-    status: AnswerStatus = AnswerStatus.Unselected,
+    isSelected: Boolean = false,
     onClick: () -> Unit = {}
 ) {
+    val backgroundColor = animateColorAsState(
+            if (isSelected) AppTheme.color.primaryVariant else AppTheme.color.surface
+        )
+    val iconColor = animateColorAsState(
+        if (isSelected) AppTheme.color.primary else AppTheme.color.body
+    )
+    val borderColor = animateColorAsState(
+        if (isSelected) Color.Unspecified else AppTheme.color.stroke
+    )
     val shape = RoundedCornerShape(16.dp)
-
     Row(
         modifier = modifier
             .fillMaxWidth()
             .background(
-                color = status.backgroundColor(),
+                color = backgroundColor.value,
                 shape = shape,
             )
             .border(
-                color = status.borderColor(),
+                color = borderColor.value,
                 width = 1.dp,
                 shape = shape,
             )
             .clip(shape)
-            .clickable(enabled = isClickable,onClick = onClick)
+            .clickable(onClick = onClick)
             .padding(vertical = 16.dp, horizontal = 12.dp),
         verticalAlignment = Alignment.CenterVertically,
     ) {
-
+        Icon(
+            painter = painterResource(id = trailingIcon),
+            contentDescription = null,
+            tint = iconColor.value
+        )
         Text(
             text = text,
             color = AppTheme.color.body,
@@ -61,42 +76,30 @@ fun AnswerSelectionItem(
                 .weight(1f)
                 .padding(start = 8.dp)
         )
-        when (status) {
-            AnswerStatus.Unselected -> {
-                RadioButton(
-                    state = RadioState.Default,
-                )
-            }
-
-            AnswerStatus.Correct, AnswerStatus.Wrong -> {
-                Icon(
-                    painter = painterResource(status.icon),
-                    contentDescription = null,
-                    tint = status.borderColor(),
-                )
-            }
-        }
+        RadioButton(
+            state = if (isSelected) RadioState.Selected else RadioState.Default,
+            onClick = onClick,
+            modifier = Modifier,
+        )
     }
 }
 
 @Composable
 @ThemeAndLocalePreviews
-private fun LanguageSelectionItemPreview() {
+private fun SelectionFieldPreview() {
     AflamiTheme {
         Column(
-            verticalArrangement = Arrangement.spacedBy(8.dp)
+            verticalArrangement = Arrangement.spacedBy(8.dp),
         ) {
-            AnswerSelectionItem(
+            ThemeSelectionItem(
                 text = stringResource(R.string.light),
-                status = AnswerStatus.Unselected,
+                isSelected = true,
+                trailingIcon = com.amsterdam.designsystem.R.drawable.ic_moon
             )
-            AnswerSelectionItem(
-                text = stringResource(R.string.light),
-                status = AnswerStatus.Correct,
-            )
-            AnswerSelectionItem(
-                text = stringResource(R.string.light),
-                status = AnswerStatus.Wrong,
+            ThemeSelectionItem(
+                text = stringResource(R.string.dark),
+                isSelected = false,
+                trailingIcon = com.amsterdam.designsystem.R.drawable.ic_sun
             )
         }
     }
