+++ ui/src/main/java/com/amsterdam/ui/screens/search/keywordSearch/sections/RecentSearchItem.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.ui.components
+package com.amsterdam.ui.screens.search.keywordSearch.sections
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
