+++ ui/src/main/java/com/amsterdam/ui/screens/search/keywordSearch/sections/RecentSearchesSection.kt
@@ -25,7 +25,6 @@ import com.amsterdam.designsystem.components.LoadingContainer
 import com.amsterdam.designsystem.components.Text
 import com.amsterdam.designsystem.components.divider.HorizontalDivider
 import com.amsterdam.designsystem.theme.AppTheme
-import com.amsterdam.ui.components.RecentSearchItem
 import com.amsterdam.viewmodel.search.keywordSearch.SearchUiState
 
 @Composable
