+++ ui/src/main/java/com/amsterdam/ui/screens/search/keywordSearch/sections/globalSearchHub/GlobalSearchHubUI.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.ui.components.globalSearchHub
+package com.amsterdam.ui.screens.search.keywordSearch.sections.globalSearchHub
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
