+++ ui/src/main/java/com/amsterdam/ui/screens/seriesDetails/component/moreTvShowLikeSection.kt
@@ -12,7 +12,7 @@ import com.amsterdam.designsystem.components.ImageErrorIndicator
 import com.amsterdam.designsystem.components.ImageLoadingIndicator
 import com.amsterdam.imageviewer.ui.SafeImageView
 import com.amsterdam.ui.application.LocalRestrictionLevel
-import com.amsterdam.ui.components.UpcomingCard
+import com.amsterdam.ui.components.movieAndTvShowDetails.UpcomingCard
 import com.amsterdam.ui.components.adaptiveGrid
 import com.amsterdam.ui.components.EmptyStateText
 import com.amsterdam.ui.utils.toSafetyLevel
